update babel 7 on demo
