+++ demo/src/index.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import './assets/styles/main.scss';
 import navConfig from './config/main-nav';
 import Root from './Root';
@@ -7,4 +8,4 @@ import routes from './routes';
 import store from './store';
 
 const MOUNT_NODE = document.getElementById('root');
-render(<Root store={store} routes={routes} navConfig={navConfig} />, MOUNT_NODE);
+render(<Router><Root store={store} routes={routes} navConfig={navConfig} /></Router>, MOUNT_NODE);
